Decouple DialogTrees from Editor
Scripts stored inside the Editor folder are not added into the build, and therefore cannot be accessed by Scripts that would be used in a game. In order to enable DialogTrees to be properly accessible, I have moved them and all related class to outside the Editor folder, and decoupled them from any Editor components.